infrastructure for color hbonds type

+++ Jmol/src/org/openscience/jmol/viewer/managers/ModelManager.java
@@ -356,7 +356,7 @@ public String getAtomInfoChime(int i) {
     return frame.bonds[i].getRadius();
   }
 
-  public byte getBondOrder(int i) {
+  public short getBondOrder(int i) {
     return frame.bonds[i].getOrder();
   }
